refactor: 흰디의 옷장 관련 파일 closet으로 명칭 변경 및 Auth 추가
