+++ src/main/java/com/themore/moamoatown/closet/dto/GetProfileInternalDTO.java
@@ -1,4 +1,4 @@
-package com.themore.moamoatown.coordi.dto;
+package com.themore.moamoatown.closet.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,15 +6,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * 프로필 업데이트 Request DTO
+ * 프로필 조회 내부 DTO
  * @author 임원정
- * @since 2024.08.25
+ * @since 2024.08.28
  * @version 1.0
  *
  * <pre>
  * 수정일        수정자        수정내용
  * ----------  --------    ---------------------------
- * 2024.08.25  	임원정        최초 생성
+ * 2024.08.28  	임원정        최초 생성
  * </pre>
  */
 
@@ -22,6 +22,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class UpdateProfileRequestDTO {
-    private String profile; // base64 인코딩된 String
+public class GetProfileInternalDTO {
+    byte[] profile; //BLOB 타입
 }
