+++ src/main/java/com/themore/moamoatown/closet/dto/MyClothesResponseDTO.java
@@ -1,4 +1,4 @@
-package com.themore.moamoatown.coordi.dto;
+package com.themore.moamoatown.closet.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,15 +6,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * 프로필 조회 내부 DTO
+ * 내 옷장 옷 리스트 Response DTO
  * @author 임원정
- * @since 2024.08.28
+ * @since 2024.08.25
  * @version 1.0
  *
  * <pre>
  * 수정일        수정자        수정내용
  * ----------  --------    ---------------------------
- * 2024.08.28  	임원정        최초 생성
+ * 2024.08.25  	임원정        최초 생성
  * </pre>
  */
 
@@ -22,6 +22,10 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class GetProfileInternalDTO {
-    byte[] profile; //BLOB 타입
+public class MyClothesResponseDTO {
+    private Long clothId;
+    private String brand;
+    private String name;
+    private Long type;
+    private String imgUrl;
 }
