+++ src/main/resources/com/themore/moamoatown/closet/mapper/ClosetMapper.xml
@@ -14,7 +14,7 @@
  * 2024.08.28   임원정       selectProfileByMemberId 추가
  * </pre>
 -->
-<mapper namespace="com.themore.moamoatown.coordi.mapper.CoordiMapper">
+<mapper namespace="com.themore.moamoatown.closet.mapper.ClosetMapper">
     <!-- 내 옷장의 옷 가져오기 -->
     <select id="selectClothesByMemberId" resultType="MyClothesResponseDTO">
         <![CDATA[
